Rework the test suite to use testing_models.
Mostly, this was to silence the warnings about non-collected tests, because
models beginning with the word "Test" were considered to be candidates for
collecting, but were not in fact test. This wasn't breaking anything, but it
was causing far too much noise in testing (an environment which has warnings
turned on).