Funeral for multilingual support.
I'll explain my rationale in the documentation, but this is not a grudge
against multilingual as such; it was the fact you had more than one possible
representation of a page, and I am against it. This code was a persistent
source of bugs, and it rarely got used so it was a largely unnecessary
complication in 99% of cases.