Don't let TinyMCE run in to its label.
Django's CSS doesn't allow us to reasonably make it not full width without
hard-coding the width of the label in custom CSS (which means it will
never work with other admin skins.

Instead, let it full-width. This is normally what we want for HTML
content as it is probably the most important thing on the page. But we
can `clear` its label like it's Some Previous Year.