Add regression tests.
Make sure that `Page.children` does not invalidate prefetches, and make sure
that `Page.get_absolute_url()` does not do anything slow.

As a consequence, setting `Page.cached_parent` turns out to be unnecessary,
so it can go! On the downside, it means *one* of the possible ways I thought
the page tree would randomly get garbaged is not real.

While we're here, we can simplify `Page.children` too.