Get some massive speedups in render_navigation with prefetch_related.
With multilingual support pulled out, this allows us to use
prefetch_related. We can have a depth at which we *think* you might want
to prefetch children, but we don't know that for a fact. So, allow this
to be specified per-request.

While we're here, add a regression test to ensure that it stays fast
forever.