Test rework: move all tests into a single test folder.
There was no point in having apps organised per app, since we never run
tests for a single app, and even if I did, it makes it harder to find
individual tests.

It means we can't use relative imports anymore, which is a good thing.

While I'm here, remove a couple of pointless tests. Importing urlconfs
and asserting that the number of URLs is the same as some fixed number
is not asserting any useful behaviour.