Add an image filter to media admin.
This is actually part of the rationalisation of "filter media library
changelist to images only" that is part of fixing #21, but it's a neat
feature to have by itself.

But WHILE WE'RE THERE, let's have one central definition of what file
extensions mean "image", and generate the rest from there. This itself,
as a side-effect, means that ImageRefField now supports .webp.

And WHILE WE'RE THERE, fix up the tests for `uncms.media.filetypes`. They
can all be test functions, rather than using Django's TestCase.