[Fix #12] Break out image editor into a separate form in the admin.
Loading megabytes of JS isn't necessary most of the time, so don't load it
all of the time.

While we're here, simplify everything. We don't need to take the features
from the settings (which no longer exist). We don't need to make the
maximum width of the editor configurable, etc.