Simplify used_on - because of get_fieldsets it is not shown if `obj == None`.
That fixes the last uncovered branch in media/admin.py (towards #17).

While we're there, don't add usages of a file that do not have an admin_url.
That'll avoid showing an empty string (rather than "no usage" in cases where
a file has an m2m relationship with something, i.e. it has a label.