Move test for page.auth_required further up.
Theoretically, if a subview of a page had a side-effect, the side-effect
could happen AND they would be redirected to the login page. If anyone was
gating real functionality on this, it'd be a security hole. I don't think
it ever *was* used for that.

While we're there, `request` is never falsey (we depend on it not being
because we're looking at `request.path`), so there's no need to test
for that.