Remove last_modified from the page list - it's extremely inefficient.
There's a case for making this a first-class field in the future, because
it is useful. But even a 10x10 page tree could result in 100 queries here.
Pity the one using RDS (or anything else that goes over a network
connection)!

While we're there, ensure that we can always get the page list with the same
number of queries regardless of our object count. This means using
select_related for the content_type as well. Add a test so this can not
regress.