[#6] Rework a whole pile of template tags.
First, for almost all of them, it is never useful to pass the parameter
you wish to output as an argument. If you want to put "blob" as your
OG description, you would not {% og_description 'Blob' %}, you would
simply put that string in your template. The only time this might be useful
is for meta robots and the Twitter card, just because the values are not
exactly the same as the output.

Secondly, og/twitter images fell back to "logo", when a logo almost never
makes a good OpenGraph image.

Thirdly, all of this was just three times as complicated as it needs to be.
Simplify it.

Finally, simplify the tests (#22). We have pytest and factories and all the
other fun. This brings our pages template tag coverage to almost 100%, and
actual 100% for all the shared code.