Re-write the old regex-based HTML processor.
Parsing HTML with regular expressions was...a decision. It is a surprise that
it worked consistently at all, let alone worked reliably, but it did. But we
can do much better with much less code given we have BeautifulSoup.

We don't expand permalinks to objects anymore. The permalinks module is going
to be retired. We do however want to have image references - we can special
case that one for now, and keep compatibility with content from the site's
ancestors (as far as I know they all used the special case of '/r/XX-YYY/').

Fixes #25.