Replace dependence on permalinks with a media file redirect view.
Permalinks were only in use internally to make a "permanent" link to a
media file in the HTML editor. Because of the issues detailed in #18
(basically making every object with `get_absolute_url` public), we want
to retire this eventually.

So instead, for its *internal* use, we can just have a tiny redirect view
instead, which redirects to a media file, if you are a staff user who
has the `media.view_file` permission.

This permits retiring permalinks entirely.