Remove the "add random sizes" element of render_multiformat_image.
This *forces* setting the "sizes" attribute of a <picture>, which is really
not for the lazy. Instead, just output the original format and WebP. That
will do. If you want to display different images at different breakpoints,
use it more than once and `display: none` or `display: block` them with
CSS.

Browsers made a mistake requiring "sizes" when width descriptors are given,
because the browser knows the width an image needs to be rendered at! It
can't prefetch it before it has loaded the CSS, but that is a tiny
optimisation. The downside of requiring "sizes" is that you need to
keep your "sizes" in your HTML template and your media queries in CSS
in sync, and ain't nobody got time for that.