[Fix #34] Add system checks to fail early if required settings are not there.
Previously, the pages & publication middleware were required, but would fail
late if they were not present. And the context processor that made 'request'
available were required to do anything useful.

So, add a system check to ensure that they are present. It will error on the
middleware and missing `django.template.context_processors.request` context
processor. It will also warn if `uncms.pages.context_processors.pages` is
missing; this is not absolutely required, but referring to "pages" rather
than "request.pages" feels nicer.