Verify PUBLICATION_MIDDLEWARE_EXCLUDE_URLS is sane.
While (in the UnCMS era) it now has sensible defaults, it might be overridden
per project, and if the admin was moved to another path *without* updating
this setting, it would cause publication middleware to not work in the
admin, which would at the least have resulted in offline objects not
appearing, and could have resulted in page tree breakage at worst.

Thus, verify that 1) it excludes your admin, if you have the admin
application installed 2) that this setting is a list or tuple, because
strings are iterable too and that would also cause silent breakage of the
admin.