Replace the HTML editor with Trumbowyg.
TinyMCE needed replacing for a few reasons:

* It was big. TinyMCE does a lot of things. It does everything imaginable.
  It was comprehensive enough to author full documents. That also means it
  does far more than is actually needed from an HTML editor on most sites,
  where the use case is "add some formatting to text on part of a page".
  Trumbowyg works great in this role. And it's about a fifth of the size
  *unminified* compared to TinyMCE *minified*.
* TinyMCE seems to be moving moving towards a "nerfed free version plus paid
  upgrades" licensing model, and also pushes cloud-hosted rather heavily.
  That is both understandable and entirely their prerogative. But that makes
  it a risky future to bet on.
* It was an old version of TinyMCE, so everything would have needed a
  re-write in any case; moving to a different editor was no extra effort.

This fixes #14, because while we still have a pile of vendored JS, it's a
pile of vendored JS of a fraction of the size, which is not a minified blob
anymore. The original plan in that ticket was to replace it with a CDN-hosted
version, but this coincided with my Internet going out for several days.
That soured me on using a CDN, because it was a reminder that I won't always
have an Internet connection while developing.

So, while I'm there...

The old insert-image-from-library tool was never great. It loaded Django's
`media.File` admin in an iframe and added some click handlers to pass
messages up to the parent window. This was...a solution, but it had a way
of breaking from time to time. It also required weakening X-Frame-Options
and/or the site's CSP, which we should not really need to do, even for a
single view. And its UI was confusing; there was a confusion between "hit
OK on dialog" and "submit image upload form".

Instead, there is a real media library plugin now, with a standalone UI
that should not break in between Django upgrades. Image uploading is
handled by Trumbowyg's own image uploader. That fixes #35.

The new file browser is also progress towards #8 - while the old file
browse/upload implementation never had to use inline scripts, it did,
and the new one does not.

Also, while we're here, `FileAdmin.remote_view` has been retired. The idea
of "upload a file from a remote URL" is a good one and this may even return
in future, but it was limited to overwriting an *existing* file, which is
much less useful. Removing this means the `requests` dependency can be
dropped, perhaps only in the short term.