Add new redirects app.
Almost every site will require something like this, and especially sites that
existed under some other CMS in the past. Django comes with its own, but it's
fairly minimally designed.

* It does not support changing the type of redirect per redirect; either all
  redirects are permanent, or none of them are.
* Its admin isn't very good (admittedly, fixable project-locally).
* It doesn't do basic validation like "does this redirect to itself"
* It's fiddly because old paths have to be entered as paths and not as full
  URLs (which can be normalised to paths)
* It depends on the `sites` framework.

None of which I want, and as a bonus minor feature (but one that when you need
it, you really need it):

* It doesn't support regex paths.

Thus, this new app. It allows pasting full URLs as the redirect URLs. It
has CSV import. It supports that regex thing. It's generally an easier-to-use
alternative than the Django one.