Fix media list layout for very long titles.
The default title for a file is generated from the filename on upload.
Sometimes, these titles contain long strings without spaces, and in these
cases, it would cause the fancy grid list to break by making some cells
much larger.

Fix this by adding a CSS rule to allow breaking anywhere, including in the
middle of a word.