[#13] Retire the in-browser image editor.
It never worked very well, it was broken for at least half the use cases
(you were likely to exceed DATA_UPLOAD_MAX_MEMORY_SIZE for photos from
modern cameras), and it was also a giant pile of vendored JS that I was
never going to update. Maybe a simpler version providing the basic
editing operations will come another day, but that is not now.