Add a much more robust locking mechanism for Page saves.
Previously, it was using select_for_update. That actually was a half
decent locking mechanism, I think, but it only guarded against concurrent
saves and didn't do anything for inserts.

Instead, let's have a global table lock. We can't actually do that
*as such* portably across databases. But we can do that with a singleton
in a dedicated table; sprinkle on some select_for_update with that and
you have a table lock that works across databases.

This should really, truly solve the problem of the page tree getting
trashed. That hasn't been seen since UnCMS came about, but it is good
to rule out concurrent requests ever being able to cause that.