[Fix #5] Fix tests on MySQL.
A lack of max_length on a test model didn't work with MySQL; only SQLite
and PostgreSQL support that.

That means I can officially call it supported. There's no reason to think
that won't work because we have 100% coverage and anything that was going
to explode because of unsupported features would have done. So add that
to the documentation.